rename: 단순 데이터 조회, 재사용 되는 특징에 따라 entities 폴더로 이동
